server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
fal-client==0.4.2
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.0
websockets==12.0
Pillow==10.3.0
